test convolving pgopher data
