+++ src/test.py
@@ -1,117 +1,79 @@
 # module test
-'''
-Model equilibrium tranmission using the Beer-Lambert law.
-'''
 
 import numpy as np
 import pandas as pd
-import matplotlib.pyplot as plt
-import scienceplots  # pylint: disable = unused-import
+from scipy.special import wofz # pylint: disable=no-name-in-module
 
 import constants as cn
+from simtype import SimType
+from molecule import Molecule
+from simulation import Simulation
 
-plt.style.use(['science', 'grid'])
+def plot_samp(samp_file: str):
+    sample_data = pd.read_csv(f'../data/samples/{samp_file}.csv')
 
+    wavenumbers = sample_data['wavenumbers'].to_numpy()
+    intensities = sample_data['intensities'].to_numpy()
+    intensities /= intensities.max()
 
-def calc_density_cm(pressure_torr, temp):
-    pressure_pa = pressure_torr * 101325 / 760
+    return wavenumbers, intensities
 
-    return pressure_pa / (cn.BOLTZ * temp) / 1e6
+def convolve_brod(sim: Simulation, wavenumbers_line: np.ndarray, intensities_line: np.ndarray,
+                  wavenumbers_conv: np.ndarray) -> np.ndarray:
+    intensities_conv = np.zeros_like(wavenumbers_conv)
+    natural, collide = broadening_params(sim)
 
+    for _, (wave, intn) in enumerate(zip(wavenumbers_line, intensities_line)):
+        intensities_conv += intn * broadening_fn(sim, wavenumbers_conv, wave, natural, collide)
 
-vib_qn_up_max  = 3
-vib_qn_lo_list = np.arange(0, 19)
-length         = 1                                      # [cm]
-temperature    = 3000                                   # [K]
-pressure       = 1                                      # [Torr]
-density        = calc_density_cm(pressure, temperature) # [cm^-3]
-cross_section  = 14.2e-18                               # [cm^2]
+    return intensities_conv
 
-# CS from: Absorption cross section of molecular oxygen, Lu, Chen
-# N * L           ~ 10^15  [cm^-2]
-# FC * CS         ~ 10^-18 [cm^2]
-# FC * CS * N * L ~ 10^-3  [-]
-# alpha = FC * CS * N
-# I/I_0 = 1 - alpha * L
+def broadening_fn(sim: Simulation, convolved_wavenumbers: np.ndarray, wavenumber_peak: float,
+                  natural: float, collide: float) -> np.ndarray:
 
-print(f'Total number density:            {density:10.4e} [cm^-3]')
-print(f'Cross section at 295K, 143.26nm: {cross_section:10.4e} [cm^2]')
+    doppler = (wavenumber_peak *
+               np.sqrt(cn.BOLTZ * sim.temp / (sim.molecule.molecular_mass * (cn.LIGHT / 1e2)**2)))
 
-df      = pd.read_csv('../data/molecular_constants/o2.csv', index_col=0)
-consts  = df.loc['x3sg']
-fc_data = np.loadtxt('../data/franck-condon/o2.csv', delimiter=',')
+    prediss = 0.1
 
+    gauss = doppler
+    loren = natural + collide + prediss
 
-def vibrational_term(vib_qn_lo):
-    return (consts['w_e']   * (vib_qn_lo + 0.5)    -
-            consts['we_xe'] * (vib_qn_lo + 0.5)**2 +
-            consts['we_ye'] * (vib_qn_lo + 0.5)**3 +
-            consts['we_ze'] * (vib_qn_lo + 0.5)**4)
+    fadd = ((convolved_wavenumbers - wavenumber_peak) + 1j * loren) / (gauss * np.sqrt(2))
 
+    return np.real(wofz(fadd)) / (gauss * np.sqrt(2 * np.pi))
 
-total_partition = (np.exp(-vibrational_term(vib_qn_lo_list) *
-                          cn.PLANC * cn.LIGHT / (cn.BOLTZ * temperature))).sum()
+def broadening_params(sim: Simulation) -> tuple[float, float]:
+    natural = (sim.state_lo.cross_section**2 *
+               np.sqrt(8 / (np.pi * sim.molecule.reduced_mass * cn.BOLTZ * sim.temp)) / 4)
 
+    collide = ((sim.pres * 10) * sim.state_lo.cross_section**2 *
+               np.sqrt(8 / (np.pi * sim.molecule.reduced_mass * cn.BOLTZ * sim.temp)) / 2)
 
-def number_density(vib_qn_lo):
-    return (density * np.exp(-vibrational_term(vib_qn_lo) * cn.PLANC * cn.LIGHT /
-                             (cn.BOLTZ * temperature)) / total_partition)
+    return natural, collide
 
+temp: float = 300.0
+pres: float = 101325.0
 
-x = np.linspace(vib_qn_lo_list.min(), vib_qn_lo_list.max(), 1000)
+mol_o2 = Molecule('o2', 'o', 'o')
 
+bands_sim = [(2, 0)]
 
-def plot_text(func):
-    for level in vib_qn_lo_list:
-        plt.text(level + 0.1, func(level), f'{level}')
+o2_sim = Simulation(mol_o2, temp, pres, np.arange(0, 36), 'b3su', 'x3sg', bands_sim,
+                    SimType.ABSORPTION)
 
+wns, ins = plot_samp('pgopher')
 
-plt.plot(vib_qn_lo_list, number_density(vib_qn_lo_list), 'o')
-plt.plot(x, number_density(x), label='Boltzmann Distribution')
-plot_text(number_density)
-plt.xlabel('Ground State Vibrational Level, $v\'\'$ [-]')
-plt.ylabel('Vibrational Number Density, $N_v$ [-]')
-plt.legend()
-plt.show()
+cwns = np.linspace(wns.min(), wns.max(), 10000)
+cins = convolve_brod(o2_sim, wns, ins, cwns)
+cins /= cins.max()
 
+def wavenum_to_wavelen(x):
+    x             = np.array(x, float)
+    near_zero     = np.isclose(x, 0)
+    x[near_zero]  = np.inf
+    x[~near_zero] = 1 / x[~near_zero]
 
-def get_fc(vib_qn_lo):
-    return fc_data[vib_qn_up][vib_qn_lo]
+    return x * 1e7
 
-
-for vib_qn_up in range(0, vib_qn_up_max):
-    plt.plot(vib_qn_lo_list, get_fc(vib_qn_lo_list), '-o', label=f'$v\' = {vib_qn_up}$')
-
-plot_text(get_fc)
-plt.xlabel('Ground State Vibrational Level, $v\'\'$ [-]')
-plt.ylabel('Franck-Condon Factor, FCF [-]')
-plt.legend()
-plt.show()
-
-
-def factor(vib_qn):
-    return get_fc(vib_qn) * cross_section * number_density(vib_qn) * length
-
-
-for vib_qn_up in range(0, vib_qn_up_max):
-    plt.plot(vib_qn_lo_list, factor(vib_qn_lo_list), '-o', label=f'$v\' = {vib_qn_up}$')
-
-plot_text(factor)
-plt.xlabel('Ground State Vibrational Level, $v\'\'$ [-]')
-plt.ylabel('Factor, $\\sigma NL$ [-]')
-plt.legend()
-plt.show()
-
-
-def transmission(vib_qn_lo):
-    return np.exp(-factor(vib_qn_lo))
-
-
-for vib_qn_up in range(0, vib_qn_up_max):
-    plt.plot(vib_qn_lo_list, transmission(vib_qn_lo_list), '-o', label=f'$v\' = {vib_qn_up}$')
-
-plot_text(transmission)
-plt.xlabel('Ground State Vibrational Level, $v\'\'$ [-]')
-plt.ylabel('Transmission, $I/I_0$ [-]')
-plt.legend()
-plt.show()
+cwls = wavenum_to_wavelen(cwns)
